+++ SistemaPasteleria/Models/ListaMaterial.cs
@@ -1,0 +1,19 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Text;
+using System.Threading.Tasks;
+using System.Data.Entity;
+using System.ComponentModel.DataAnnotations;
+
+namespace SistemaPasteleria.Models
+{
+    public class ListaMaterial
+    {
+        [Key]
+        public int IdListaMaterial { get; set; }
+        public int IdReceta { get; set; }
+        public int IdCatalogo { get; set; }
+        public double Cantidad { get; set; }
+    }
+}
