+++ SistemaPasteleria/packages.config
@@ -1,11 +1,14 @@
 ﻿<?xml version="1.0" encoding="utf-8"?>
 <packages>
+  <package id="CsvHelper" version="24.0.1" targetFramework="net472" />
   <package id="EntityFramework" version="6.4.4" targetFramework="net472" />
   <package id="EntityFramework6.Npgsql" version="6.4.1" targetFramework="net472" />
   <package id="Microsoft.Bcl.AsyncInterfaces" version="1.1.0" targetFramework="net472" />
+  <package id="Microsoft.Bcl.HashCode" version="1.1.0" targetFramework="net472" />
+  <package id="Microsoft.CSharp" version="4.5.0" targetFramework="net472" />
   <package id="Npgsql" version="5.0.3" targetFramework="net472" />
   <package id="System.Buffers" version="4.5.1" targetFramework="net472" />
-  <package id="System.Memory" version="4.5.3" targetFramework="net472" />
+  <package id="System.Memory" version="4.5.4" targetFramework="net472" />
   <package id="System.Numerics.Vectors" version="4.5.0" targetFramework="net472" />
   <package id="System.Runtime.CompilerServices.Unsafe" version="4.6.0" targetFramework="net472" />
   <package id="System.Text.Encodings.Web" version="4.6.0" targetFramework="net472" />
